requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
requests==2.31.0
python-dotenv==1.0.0
woocommerce==3.0.0
selenium==4.15.2
beautifulsoup4==4.12.2
openpyxl==3.1.2
sqlalchemy==2.0.23
python-dateutil==2.8.2
scipy==1.11.4

